requirements.txt
vk_api==11.9.0
facebook-sdk==3.1.0